LayoutTests/crypto/subtle/aes-cbc-import-key-unwrap-pkcs8-key-expected.txt

Test unwrapping a PKCS8 key with AES-CBC using an imported key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new Uint8Array(unwrappedKey) is pkcs8Key
PASS successfullyParsed is true

TEST COMPLETE
